Add script to fetch and display posts
This commit adds a new script to the project that fetches data from the "https://jsonplaceholder.typicode.com/posts" API and displays the fetched posts on the webpage. The script listens for a click event on the button with the id "traerPosts" and fetches the data asynchronously. The fetched posts are then dynamically added to the div with the id "post-data" as a list of items.

The commit also includes the addition of a new HTML file, "index.html", which contains the necessary markup for the webpage. The script file, "script.js", is linked to this HTML file.

This commit addresses the need to fetch and display posts on the webpage.